move transfer function to bottom and clean up code
