+++ .gitignore
@@ -6,3 +6,4 @@
 .DS_Store
 .venv
 frontend/node_modules
+frontend/dist
