+++ frontend/tsconfig.tsbuildinfo
@@ -1,1 +1,1 @@
-{"root":["./src/app.tsx","./src/main.tsx","./src/theme.ts","./src/vite-env.d.ts","./src/api/circuitapi.ts","./src/components/analysis/bodeplotpanel.tsx","./src/components/analysis/loopgainpanel.tsx","./src/components/analysis/schematicpanel.tsx","./src/components/analysis/stabilitypanel.tsx","./src/components/analysis/transferfunctionpanel.tsx","./src/components/landing/landingpage.tsx","./src/components/parameters/parameterpanel.tsx","./src/components/sfg/edgeeditmodal.tsx","./src/components/sfg/sfgembed.tsx","./src/components/sfg/sfgtoolbar.tsx","./src/components/sfg/sfgviewer.tsx","./src/components/sfg/svgoverlay.tsx","./src/components/sidebar/analysissidebar.tsx","./src/context/circuitcontext.tsx","./src/pages/sfgapppage.tsx","./src/types/cytoscape-dagre.d.ts","./src/types/index.ts","./src/utils/cytoscapestyles.ts","./src/utils/edgecurves.ts","./src/utils/formatting.ts"],"version":"5.9.3"}
+{"root":["./src/app.tsx","./src/main.tsx","./src/theme.ts","./src/vite-env.d.ts","./src/api/circuitapi.ts","./src/components/analysis/bodechart.tsx","./src/components/analysis/loopgainpanel.tsx","./src/components/analysis/parameterpanel.tsx","./src/components/analysis/schematicpanel.tsx","./src/components/analysis/stabilitypanel.tsx","./src/components/analysis/transferfunctionpanel.tsx","./src/components/landing/landingpage.tsx","./src/components/parameters/parameterpanel.tsx","./src/components/sfg/edgeeditmodal.tsx","./src/components/sfg/sfgembed.tsx","./src/components/sfg/sfgtoolbar.tsx","./src/components/sfg/sfgviewer.tsx","./src/components/sfg/svgoverlay.tsx","./src/components/sidebar/analysissidebar.tsx","./src/components/sidebar/sidebarstyles.ts","./src/context/circuitcontext.tsx","./src/pages/sfgapppage.tsx","./src/types/cytoscape-dagre.d.ts","./src/types/index.ts","./src/utils/cytoscapestyles.ts","./src/utils/edgecurves.ts","./src/utils/formatting.ts"],"version":"5.9.3"}
